Moved file names and locations around (putting c code in .external)
